adds additional safe guard to `db:hard_reset` rake task

+++ backend/lib/tasks/reset_database.rake
@@ -12,7 +12,14 @@ namespace :db do
   desc "Drops, creates, migrates, and seeds your local database."
   task hard_reset: [:environment] do
     raise "You are not allowed to run this task in production!!!" unless Rails.env.development? || Rails.env.test?
-    
+    db_configs = YAML.load_file(Rails.root.join('config', 'database.yml'),  aliases: true)[Rails.env]
+    if !db_configs['transactional']['database'].ends_with?('dev')
+      raise "In order to hard_reset the database, the database name must end in `dev`. transactional DB name is: #{db_configs['transactional']['database']}"
+    end
+    if !db_configs['clickhouse']['database'].ends_with?('dev')
+      raise "In order to hard_reset the database, the database name must end in `dev`. clickhouse DB name is: #{db_configs['clickhouse']['database']}"
+    end
+
     puts "Dropping Postgres and Clickhouse DBs...".colorize(:yellow)
     run_command_with_success('bundle exec rake db:drop')
